backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.2
aiofiles==23.2.1
numpy==1.24.3
openai==1.10.0
reportlab==4.0.9
python-docx==1.1.0
